pyproject.toml
[build-system]
requires = [
    "hatchling>=1.6",
]
build-backend = "hatchling.build"

[project]
name = "highcharts-gantt"
description = "High-end Gantt Chart Visualization for the Python Ecosystem"
readme = "README.rst"
license-files = { paths = ["LICENSE"] }

dynamic = [
     "version"
]

authors = [
    { name = "HCP LLC", email = "support@highchartspython.com" }
]

keywords = [
    "highcharts",
    "data visualization",
    "data viz",
    "charts",
    "graphing",
    "highcharts gantt",
    "plotting",
    "gantt",
    "project management"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Manufacturing",
    "Intended Audience :: Science/Research",

    "Framework :: Jupyter",

    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Scientific/Engineering :: Information Analysis",

    "License :: Other/Proprietary License"
]

requires-python = ">= 3.10"
dependencies = [
    "highcharts-stock>=1.0.0-rc1",
    "esprima>=4.0.1",
    "validator-collection>=1.5.0",
    "requests>=2.28.1"
]

[tool.hatch.version]
path = "highcharts_gantt/__version__.py"

[project.urls]
"Homepage" = "https://www.highchartspython.com"
"Documentation" = "https://gantt-docs.highchartspython.com/en/latest/"
"Support" = "https://www.highchartspython.com/get-help"
"Source Code" = "https://github.com/highcharts-for-python/highcharts-gantt"
"History" = "https://github.com/highcharts-for-python/highcharts-gantt/blob/master/CHANGES.rst"
"Bug Tracker" = "https://github.com/highcharts-for-python/highcharts-gantt/issues"
# TODO: Add a Sponsor section
# "Sponsor" = "https://github.com/sponsors/highcharts-for-python"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.2",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=2.5.0",
    "python-dotenv>=1.0.0",
    "pytz>=2022.1",
    "Sphinx==6.1.3",
    "sphinx-rtd-theme==1.2.0",
    "sphinx-toolbox==3.4.0",
    "sphinx-tabs==3.4.1",
    "tox>=4.0.0",
    "IPython>=8.10.0",
    "pyspark>=3.3.0",
    "pandas>=1.3.3",
    "orjson>=3.7.7",
    "asana>=3.0.0",
    "monday>=1.3.0",
    "jira>=3.4.0"
]
soft = [
    "IPython>=8.10.0",
    "pyspark>=3.3.0",
    "pandas>=1.3.3",
    "orjson>=3.7.7",
    "asana>=3.0.0",
    "monday>=1.3.0",
    "jira>=3.4.0"
]
docs = [
    "Sphinx==6.1.3",
    "sphinx-rtd-theme==1.2.0",
    "sphinx-toolbox==3.4.0",
    "sphinx-tabs==3.4.1"
]